i dont know whats worth fighting for or why i have to scream
