+++ packages/synapse-interface/components/StateManagedBridge/InputContainer.tsx
@@ -1,20 +1,17 @@
-import React, { useEffect, useState, useCallback, useMemo } from 'react'
+import React, { useEffect, useState, useCallback } from 'react'
 import { useDispatch } from 'react-redux'
-import { useAccount, useNetwork } from 'wagmi'
+import { useAccount } from 'wagmi'
 
 import { initialState, updateFromValue } from '@/slices/bridge/reducer'
 import MiniMaxButton from '../buttons/MiniMaxButton'
 import { formatBigIntToString } from '@/utils/bigint/format'
 import { cleanNumberInput } from '@/utils/cleanNumberInput'
-import {
-  ConnectToNetworkButton,
-  ConnectWalletButton,
-  ConnectedIndicator,
-} from '@/components/ConnectionIndicators'
+
 import { FromChainSelector } from './FromChainSelector'
 import { FromTokenSelector } from './FromTokenSelector'
 import { useBridgeState } from '@/slices/bridge/hooks'
 import { usePortfolioState } from '@/slices/portfolio/hooks'
+import { ConnectStatusIndicator } from '@/components/bridgeSwap/ConnectStatusIndicator'
 
 export const inputRef = React.createRef<HTMLInputElement>()
 
@@ -22,16 +19,9 @@ export const InputContainer = () => {
   const { fromChainId, fromToken, fromValue } = useBridgeState()
   const [showValue, setShowValue] = useState('')
 
-  const [hasMounted, setHasMounted] = useState(false)
-
   const { balances } = usePortfolioState()
 
-  useEffect(() => {
-    setHasMounted(true)
-  }, [])
-
   const { isConnected } = useAccount()
-  const { chain } = useNetwork()
 
   const dispatch = useDispatch()
 
@@ -80,15 +70,6 @@ export const InputContainer = () => {
     )
   }, [balance, fromChainId, fromToken])
 
-  const connectedStatus = useMemo(() => {
-    if (hasMounted && !isConnected) {
-      return <ConnectWalletButton />
-    } else if (hasMounted && isConnected && fromChainId === chain.id) {
-      return <ConnectedIndicator />
-    } else if (hasMounted && isConnected && fromChainId !== chain.id) {
-      return <ConnectToNetworkButton chainId={fromChainId} />
-    }
-  }, [chain, fromChainId, isConnected, hasMounted])
 
   return (
     <div
@@ -97,7 +78,7 @@ export const InputContainer = () => {
     >
       <div className="flex items-center justify-between mb-3">
         <FromChainSelector />
-        {connectedStatus}
+        <ConnectStatusIndicator targetChainId={fromChainId} />
       </div>
       <div className="flex h-16 mb-2 space-x-2">
         <div
@@ -139,7 +120,7 @@ export const InputContainer = () => {
                   style={{ display: 'table-cell', width: '100%' }}
                 />
               </div>
-              {hasMounted && isConnected && (
+              {isConnected && (
                 <label
                   htmlFor="inputRow"
                   className="text-xs text-white transition-all duration-150 transform-gpu hover:text-opacity-70 hover:cursor-pointer"
@@ -155,10 +136,10 @@ export const InputContainer = () => {
             </div>
           </div>
           <div>
-            {hasMounted && isConnected && (
+            {isConnected && (
               <div className="m">
                 <MiniMaxButton
-                  disabled={!balance || balance === 0n ? true : false}
+                  disabled={!balance || balance === 0n}
                   onClickBalance={onMaxBalance}
                 />
               </div>
